refactor and add shared module ui
