+++ mobile/src/main/java/rs/readahead/washington/mobile/views/activity/onboarding/OnBoardLockFragment.kt
@@ -7,13 +7,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import com.hzontal.shared_ui.buttons.InformationButton
 import com.hzontal.tella_locking_ui.IS_FROM_SETTINGS
 import com.hzontal.tella_locking_ui.ui.password.SetPasswordActivity
 import com.hzontal.tella_locking_ui.ui.pattern.PatternSetActivity
 import com.hzontal.tella_locking_ui.ui.pin.SetPinActivity
 import rs.readahead.washington.mobile.R
 import rs.readahead.washington.mobile.views.base_ui.BaseFragment
-import rs.readahead.washington.mobile.views.custom.InformationButton
 
 class OnBoardLockFragment : BaseFragment() {
     private lateinit var lockPasswordBtn: InformationButton
