+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':shared-ui'
 include ':mobile'
 include ':tella-keys'
 include ':tella-locking-ui'
