+++ shared-ui/build.gradle
@@ -5,7 +5,7 @@ plugins {
 
 android {
     compileSdkVersion 30
-    buildToolsVersion "29.0.3"
+    buildToolsVersion "30.0.2"
 
     defaultConfig {
         minSdkVersion 17
@@ -33,28 +33,13 @@ android {
 }
 
 dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    api project(':tella-keys')
-    api project(':shared-ui')
 
-    api 'androidx.appcompat:appcompat:1.2.0'
-    api 'com.google.android.material:material:1.2.1'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation 'androidx.core:core-ktx:1.5.0'
+    implementation 'androidx.appcompat:appcompat:1.3.0'
+    implementation 'com.google.android.material:material:1.3.0'
+    api 'androidx.constraintlayout:constraintlayout:2.0.4'
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-
-    api 'androidx.constraintlayout:constraintlayout:2.0.4'
-
-    api 'androidx.appcompat:appcompat:1.2.0'
-    api 'androidx.activity:activity-ktx:1.2.0-rc01'
-    api 'androidx.fragment:fragment-ktx:1.3.0-rc01'
-
-    api "androidx.lifecycle:lifecycle-process:2.2.0"
-    api "androidx.lifecycle:lifecycle-common-java8:2.2.0"
-
-    api 'net.zetetic:android-database-sqlcipher:3.5.7@aar'
-    api 'androidx.recyclerview:recyclerview:1.1.0'
-    api "com.google.android.material:material:1.3.0"
-
-
 }
