+++ shared-ui/src/main/res/layout/layout_bottom_message.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/DialogBottomPopUpBackground"
+    android:elevation="2dp"
+    >
+
+    <TextView
+        android:id="@+id/txv_msg"
+        style="@style/DialogBottomPopUpText"
+        tools:text="MESSAGEMESSAGEMESSAGEMESSAGEMESSAGE" />
+</LinearLayout>
