+++ shared-ui/src/main/res/values/attrs.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="InformationButton">
+        <attr name="topText" format="reference|string" />
+        <attr name="topTextCenter" format="reference|string" />
+        <attr name="bottomTextCenter" format="reference|string" />
+        <attr name="bottomText" format="reference|string" />
+        <attr name="startImg" format="reference" />
+        <attr name="state_checked" format="boolean"/>
+    </declare-styleable>
+</resources>
