+++ tella-locking-ui/src/main/java/com/hzontal/tella_locking_ui/patternlock/ConfirmSetPatternActivity.java
@@ -6,8 +6,8 @@ import android.view.View;
 import androidx.core.content.ContextCompat;
 import androidx.interpolator.view.animation.FastOutLinearInInterpolator;
 
+import com.hzontal.shared_ui.utils.DialogUtils;
 import com.hzontal.tella_locking_ui.R;
-import com.hzontal.tella_locking_ui.ui.utils.DialogUtils;
 
 import java.util.ArrayList;
 import java.util.List;
