+++ tella-locking-ui/src/main/java/com/hzontal/tella_locking_ui/patternlock/SetPatternActivity.java
@@ -11,9 +11,9 @@ import android.os.Bundle;
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.core.content.ContextCompat;
 
+import com.hzontal.shared_ui.utils.DialogUtils;
 import com.hzontal.tella_locking_ui.R;
 import com.hzontal.tella_locking_ui.ui.pattern.PatternSetConfirmActivity;
-import com.hzontal.tella_locking_ui.ui.utils.DialogUtils;
 
 import java.util.ArrayList;
 import java.util.List;
