+++ tella-locking-ui/src/main/res/values/styles.xml
@@ -1,10 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<!--
-  ~ Copyright (c) 2015 Zhang Hai <Dreaming.in.Code.ZH@Gmail.com>
-  ~ All Rights Reserved.
-  -->
-
 <resources>
 
     <style name="Base.PatternView" parent="">
@@ -36,17 +31,6 @@
         <item name="android:gravity">center</item>
     </style>
 
-    <style name="DialogBottomPopUpText">
-        <item name="android:layout_width">match_parent</item>
-        <item name="android:layout_height">match_parent</item>
-        <item name="android:fontFamily">@font/open_sans</item>
-        <item name="android:layout_gravity">center_horizontal</item>
-        <item name="android:gravity">center_vertical</item>
-        <item name="android:paddingLeft">15dp</item>
-        <item name="android:paddingRight">15dp</item>
-        <item name="android:textColor">@android:color/black</item>
-        <item name="android:textSize">15sp</item>
-    </style>
 
     <style name="DialogBottomPopUpBackground">
         <item name="android:layout_width">match_parent</item>
